+++ src/components/field/field.tsx
@@ -47,6 +47,7 @@ export const Field: OverridableComponent<TypeMap> = React.forwardRef<
 
     return (
         <Component
+            data-qa="field"
             {...rest}
             ref={ref}
             className={classNames(rest.className, classes.root, {